moved api-call related files into one folder
