+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import { createMuiTheme, ThemeProvider, Theme } from "@material-ui/core/styles"
 import Nav from "./components/nav"
 import Login from "./components/login"
 import Project from "./components/project"
-import { tokenFetcher, tokenValidator } from "./lib/auth"
+import { tokenFetcher, tokenValidator } from "./lib/api/auth"
 import CssBaseline from "@material-ui/core/CssBaseline"
 import {
   BrowserRouter as Router,
